Reorg physics module. 'systems', 'components' and 'resources' modules have had their respective elements congregated into a 'static' and 'dynamic' module
