+++ src/lib.rs
@@ -160,6 +160,10 @@
 //!
 //! ## [Avian2d integration](`crate::physics`)
 //!
+//! - [Dynamic rigid bodies](`crate::physics::dynamic`) — promote particles into physics-driven
+//!   rigid body proxies and rejoin them back into the simulation
+//! - [Static rigid bodies](`crate::physics::static`) — per-chunk collision mesh generation from
+//!   marked particles
 //! - [Tagging particles as static rigid bodies](`StaticRigidBodyParticle`)
 //! - [Promoting particles to dynamic rigid bodies](`DynamicRigidBodySignal`)
 //! - [Configuring collision mesh calculation intervals](DirtyChunkUpdateInterval)
