+++ src/prelude.rs
@@ -17,8 +17,8 @@ pub use crate::render::{
     extract_chunk_image, ChunkEffectApp, ChunkEffectLayer, ChunkEffectMaterial,
     ChunkRenderingConfig, ChunkRenderingPlugin, ColorAssignment, ColorGradient, ColorIndex,
     ColorProfile, ColorRng, ColorSource, DefaultChunkEffectMaterial, FallingSandRenderPlugin,
-    ForceColor, Palette, ParticleColor, RenderingSystems, TextureSource,
-    WithColor, WorldColorTexture, WorldEffectTexture,
+    ForceColor, Palette, ParticleColor, RenderingSystems, TextureSource, WithColor,
+    WorldColorTexture, WorldEffectTexture,
 };
 
 #[cfg(feature = "movement")]
@@ -44,7 +44,8 @@ pub use crate::reactions::{
 #[cfg(feature = "physics")]
 pub use crate::physics::{
     DirtyChunkUpdateInterval, DouglasPeuckerEpsilon, DynamicRigidBodyProxy,
-    DynamicRigidBodySignal, FallingSandPhysicsPlugin, StaticRigidBodyParticle, SuspendedParticle,
+    EmitDynamicRigidBodyParticleSignal, FallingSandPhysicsPlugin, StaticRigidBodyParticle,
+    SuspendedParticle,
 };
 
 #[cfg(feature = "persistence")]
